amdgpu: Allow creating userptr BO's from files
When creating BOs from a userptr, AMDGPU_GEM_USERPTR_ANONONLY was
specified which will cause validation to fail if the mapping is backed
by a file.

This is an issue when trying to import shm-backed buffers, as might
often be the case for Wayland clients. Remove this restriction to allow
the use of shm-backed buffers.