include: split out UTS specific UAPI headers
The UAPI depends on the actual UTS. FreeBSD happens to be close enough to
Linux, so we can use the same header (which is copied over from Linux uapi
includes), but others (eg. SunOS/Illumos) at least need some patching.

To get rid of unpleasant workarounds like UTS specific semi-forks, it's better
split the UAPI headers per UTS.

For Linux (and those close enough to it), the only practical change is the
drm.h file being moved to some different directory in this source tree.

For start, the SunOS one is initially copied over from the Linux's one and
adapted by subsequent patches.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>